Tasks 83 - 92 Complete
I completed all the required tasks for the assignment and have modified the header. I also adjusted the prewritten code slightly to better fit the way I like to organize my code. Also I made sure to leave a comment above the different sections for each different question that we had to answer.